+++ trl/experimental/bco/bco_trainer.py
@@ -1435,9 +1435,9 @@ class BCOTrainer(_BaseTrainer):
         # is counted even when the other ranks are finite. This trainer logs through `store_metrics`. Evaluation goes
         # through `prediction_step`, which never calls this method, so only training reaches here and the rate belongs
         # under training.
-        # Cast before the reduction: a low-precision dtype such as `float8_e5m2` has no `mean` kernel, and the
-        # cast leaves the finiteness of every other dtype unchanged.
-        nonfinite = self.accelerator.gather((~torch.isfinite(loss.detach().float().mean())).float())
+        # Widen before the reduction: a low-precision dtype such as `float8_e5m2` has no `mean` kernel, and
+        # narrowing instead would report a finite `float64` loss above `float32`'s maximum as non-finite.
+        nonfinite = self.accelerator.gather((~torch.isfinite(loss.detach().double().mean())).float())
         self.store_metrics({"frac_nonfinite_loss": nonfinite.mean().item()}, train_eval="train")
         if nonfinite.any():
             logger.warning_once(
